Propper Energy UI Update
The issue was resolved - turned out to be 'user error' on my end - so here is, hopefully, the full (if rudimentary) setup. Again: If you leave energy-use at 0, the number and slider won't change. (I set it to 0.1 during testing. That's still pretty strict, but lets you test it out)