dragAnimation v0.1
I've got versions of the script with more work than this (this is just treating the arms like legs) but this is the last version that works - after this, I have enough half-finished changes that it stops working. More for getting the script and model into the game so I can edit them later to get the real version done.